'add header and bg img '

+++ Todo/views.py
@@ -1,23 +1,20 @@
-from django.shortcuts import render , redirect
 from .models import TodoModel
+from django.shortcuts import render , redirect
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 # Create your views here.
-@login_required(login_url='Home')
+@login_required(login_url='login')
 def change(request,id):
-        if request.method == "GET":
-            work = TodoModel.objects.get(id=id)
-            if request.user == work.user:
-                done = request.GET.get('done')
-                if done == 'on':
-                    work.done = True
-                    work.save()
-                return redirect('Home')
-            else:
-                return redirect('Home')
-        else:
-            pass
+    work = TodoModel.objects.get(id=id)
+    
+    if request.user == work.user:
+        work.done = True
+        work.save()
+        return redirect('Home')
+    else:
+        return HttpResponse('403')
 
-@login_required(login_url='Home')
+@login_required(login_url='login')
 def add_work(request):
     if request.method == 'GET':
         work = TodoModel()
@@ -29,7 +26,11 @@ def add_work(request):
     else:
         pass
     
+@login_required(login_url='login')    
 def delete_work(request,id):
     work = TodoModel.objects.get(id=id)
-    work.delete()
-    return redirect('Home')
+    if request.user == work.user:
+        work.delete()
+        return redirect('Home')
+    else :
+        return HttpResponse('403')
